feat(add-chart): add new chart popover
